Prompt timeline: add an overview-ruler style selectable by setting
Introduce a second rail style and let users choose via the new
sessions.promptTimeline.style setting ('compact' | 'overview').

- Extract IPromptTimelineRail as the seam the contrib consumes, and a
  shared PromptTimelineCard (hover preview) used by both styles.
- Rename the existing rail to PromptTimelinePillRail (dense pills).
- Add PromptTimelineRulerRail: the session compressed into the rail like
  the editor overview ruler — proportional marks at their transcript
  scroll positions, a scrollbar-slider you-are-here thumb, one green
  "changed code" signal (editorOverviewRuler.addedForeground), focusBorder
  active, >=24px hit targets. Detail stays in the shared hover card.
- Model exposes getScrollLayout()/onDidChangeScrollLayout for proportional
  positions (per-request offsets via currentRenderedHeight + scrollTop).
- Contrib picks the rail class reactively and reuses the enablement swap
  pattern, wiring setScrollLayout only for rails that support it.

All colors come from real theme tokens. Validated: typecheck, eslint,
layers, hygiene, unit tests (12 passing).
