+++ src/vs/sessions/contrib/promptTimeline/browser/promptTimelineWidgetContrib.ts
@@ -11,15 +11,17 @@ import { IInstantiationService } from '../../../../platform/instantiation/common
 import { IChatWidget } from '../../../../workbench/contrib/chat/browser/chat.js';
 import { IChatWidgetContrib, ChatWidget } from '../../../../workbench/contrib/chat/browser/widget/chatWidget.js';
 import { ChatAgentLocation } from '../../../../workbench/contrib/chat/common/constants.js';
-import { MIN_PROMPTS, PROMPT_TIMELINE_CONTRIB_ID, PROMPT_TIMELINE_ENABLED_SETTING } from '../common/promptTimeline.js';
+import { MIN_PROMPTS, PROMPT_TIMELINE_CONTRIB_ID, PROMPT_TIMELINE_ENABLED_SETTING, PROMPT_TIMELINE_STYLE_SETTING, PromptTimelineStyle } from '../common/promptTimeline.js';
 import { PromptTimelineModel, PromptEntry, PromptTick } from './promptTimelineModel.js';
-import { PromptTimelineRail } from './promptTimelineRail.js';
+import { IPromptTimelineRail } from './promptTimelineRail.js';
+import { PromptTimelinePillRail } from './promptTimelinePillRail.js';
+import { PromptTimelineRulerRail } from './promptTimelineRulerRail.js';
 
 /**
- * Per-widget contribution that overlays the {@link PromptTimelineRail} on the
- * chat transcript and exposes a navigation API for keyboard-driven commands.
- * The rail is created only while `sessions.promptTimeline.enabled` is set, and
- * is torn down (and re-created) as the setting is toggled.
+ * Per-widget contribution that overlays a prompt timeline rail on the chat
+ * transcript and exposes a navigation API for keyboard-driven commands. The rail
+ * exists only while `sessions.promptTimeline.enabled` is set, and is torn down
+ * and re-created when the enablement or `sessions.promptTimeline.style` changes.
  */
 export class PromptTimelineWidgetContrib extends Disposable implements IChatWidgetContrib {
 
@@ -27,11 +29,11 @@ export class PromptTimelineWidgetContrib extends Disposable implements IChatWidg
 	readonly id = PromptTimelineWidgetContrib.ID;
 
 	private _model: PromptTimelineModel | undefined;
-	private _rail: PromptTimelineRail | undefined;
+	private _rail: IPromptTimelineRail | undefined;
 
 	/** Holds the model, rail and all their wiring while the feature is enabled. */
 	private readonly _enablement = this._register(new DisposableStore());
-	private _enabled = false;
+	private _railKey: string | undefined;
 
 	constructor(
 		private readonly widget: IChatWidget,
@@ -46,32 +48,41 @@ export class PromptTimelineWidgetContrib extends Disposable implements IChatWidg
 		}
 
 		this._register(this.configurationService.onDidChangeConfiguration(e => {
-			if (e.affectsConfiguration(PROMPT_TIMELINE_ENABLED_SETTING)) {
-				this._updateEnablement();
+			if (e.affectsConfiguration(PROMPT_TIMELINE_ENABLED_SETTING) || e.affectsConfiguration(PROMPT_TIMELINE_STYLE_SETTING)) {
+				this._updateRail();
 			}
 		}));
-		this._updateEnablement();
+		this._updateRail();
 	}
 
-	private _updateEnablement(): void {
+	private _style(): PromptTimelineStyle {
+		return this.configurationService.getValue<PromptTimelineStyle>(PROMPT_TIMELINE_STYLE_SETTING) === PromptTimelineStyle.Overview
+			? PromptTimelineStyle.Overview
+			: PromptTimelineStyle.Compact;
+	}
+
+	/** Creates, disposes or swaps the rail to match the enablement + style settings. */
+	private _updateRail(): void {
 		const enabled = this.configurationService.getValue<boolean>(PROMPT_TIMELINE_ENABLED_SETTING) !== false;
-		if (enabled === this._enabled) {
+		const key = enabled ? this._style() : undefined;
+		if (key === this._railKey) {
 			return;
 		}
-		this._enabled = enabled;
-		if (enabled) {
-			this._createRail();
-		} else {
-			this._enablement.clear();
-			this._model = undefined;
-			this._rail = undefined;
+		this._railKey = key;
+		this._enablement.clear();
+		this._model = undefined;
+		this._rail = undefined;
+		if (key !== undefined) {
+			this._createRail(key);
 		}
 	}
 
-	private _createRail(): void {
+	private _createRail(style: PromptTimelineStyle): void {
 		// CONTRIBS always constructs contribs with the concrete widget.
 		const model = this._enablement.add(this.instantiationService.createInstance(PromptTimelineModel, this.widget as ChatWidget));
-		const rail = this._enablement.add(this.instantiationService.createInstance(PromptTimelineRail));
+		const rail: IPromptTimelineRail = this._enablement.add(style === PromptTimelineStyle.Overview
+			? new PromptTimelineRulerRail()
+			: new PromptTimelinePillRail());
 		this._model = model;
 		this._rail = rail;
 
@@ -94,9 +105,17 @@ export class PromptTimelineWidgetContrib extends Disposable implements IChatWidg
 		this._enablement.add(autorun(reader => {
 			rail.setActive(model.activeRequestId.read(reader));
 		}));
+
+		// The overview-ruler rail needs proportional scroll positions.
+		if (rail.setScrollLayout) {
+			this._enablement.add(autorun(reader => {
+				model.onDidChangeScrollLayout.read(reader);
+				rail.setScrollLayout!(model.getScrollLayout());
+			}));
+		}
 	}
 
-	private _mountRail(rail: PromptTimelineRail): void {
+	private _mountRail(rail: IPromptTimelineRail): void {
 		const railNode = rail.domNode;
 		const host = this.widget.domNode;
 		// Ensure the overlay resolves against the widget container.
